refractoting for better long term look
