+++ src/henrybryan/friendquiz/presentation/ForAlyssa.java
@@ -1,0 +1,24 @@
+/**
+ * 
+ */
+package henrybryan.friendquiz.presentation;
+
+import java.util.Date;
+
+import henrybryan.friendquiz.objects.TheGame;
+public class ForAlyssa {
+
+	/**
+	 * @param args
+	 */
+	public static void main(String[] args) {
+		Date myDate;
+	    TheGame game = new TheGame();
+	    game.play();
+	    myDate = new Date();
+	    System.out.println("Sefini");
+	    System.out.println(myDate);
+
+	}
+
+}
