--- src/presentation/ForAlyssa.java
@@ -1,28 +1,0 @@
-/**
- * 
- */
-package presentation;
-
-import java.util.Date;
-
-/**
- * @author Bryan Wodi <talk2kamp@gmail.com>
- *
- */
-import objects.TheGame;
-public class ForAlyssa {
-
-	/**
-	 * @param args
-	 */
-	public static void main(String[] args) {
-		Date myDate;
-	    TheGame game = new TheGame();
-	    game.play();
-	    myDate = new Date();
-	    System.out.println("Sefini");
-	    System.out.println(myDate);
-
-	}
-
-}
